libraries, css, list
Although the implementation is still broken I'm now using the libraries modules to find the parsers.
Some updates to the css parts. These are primarily preparations for adding custom classes and making the style cascading.
Implementation of ul and ol, an attempt to prevent the empty spaces added here but it doesn't work so I should revert it back.
Some cleaning up of notes etc.